feat: split components in data entery component
